Further refinement of daemon, but still incomplete
On starting it waits ten seconds (proper time TBD) to allow wifi
connection, if any. Then loops. On each iter:

if there are no netmanager reported wifi connections, wlan0 is set
to managed by NM (if it is not already), then ssids are retrieved
and if any found, we will (but don't yet) start Management mode
http server (if not running) and bring up the wifi-ap, else we
loop until we get ssids.

If on an iter wifi is connected, we will (but don't yet) set to
Operational mode by starting the Opertaional mode http server (if
not already running).

Testable now. After connecting the interfaces, you can connect to
external wifi with:
$ wifi-connect.netman, after which you see this
in journal every five seconds:

==== Start Operational Mode http server...

Then disconnect from external wifi with:
$ wifi-connect.netman-disconnect-wifi
after which you see this every five seconds

==== No network manager wifi connection
==== Start Management Mode http server...
==== Start wifi-ap AP...

Current impl needs following change: when no wifi connection, so
Management mode, first check if we have any known ssids. if yes,
check if wifi-AP is UP, and if not, put it UP. If no known ssids,
bring wifi-AP down, set wlan0 MANAGED, and get SSIDS (so on future
loop iter the IP is put UP.)