Daemon works to current expectation
When device is not connected to external wifi, wifi-ap and management
portal are up. User selects AP, provides password, and attempts to
connect. Device takes down wifi-ap and stops managment portal, then
waits to give enough time for connection success or failure. On success,
device is in Operational mode, and that server is started (although the
page does not yet exist). On failure, back to Management Mode, wifi-ap
and portal are put UP.

Technical points include:

* "wait" flag file now written and deleted by Management server (via new utils pkg functions). Flag file controls daemon loop. When present, Management server is UP, device is in Management mode, and we await user input. Daemon loops, doing nothing.  When not present, deamon loop takes appropriate action based on state.
* added convenience functions to:
  - manage/unmanage wlan0
  - start and stop management and operational servers, only if state
    actually needs changing
* SSIDs can only be retrieved when wlan0 is managed by network manager, so now when we enter Management mode, first we iterate until we get ssids, they are written to file in snap_common, then wlan0 is set to unmanaged and the wifi-ap is put UP.
* changed sleep times. For example, on first entrance of main loop, we wait 40 seconds to give time for any wifi connections to come up before we loop as normal and check state.