add wait on first loop before bringing down wifi-ap
Now and then on first install disabling wifi-ap may not worki in first
iter of daemon, I think due to a race condition with wifi-ap itself. It
may then think it is up normally (but it has no ssids so is broken)

I've added a wait before this wifi-ap disabling on the first loop of
the daemon to allow any other process (like wifi-ap itself) to reach
steady state before we take over.