Replace sleeps with tests
SetIfaceManaged now checks the state of the iface after changing the
config and loops until the state reaches the config state or the max
number of loop iters is reached. This allows removing time.Sleep()
calls elsewhere which had been used to ensure the iface had time to
reach the configured state.

ConnectAp() now checks if a connection has been made, also looping
as above with a max number of iterations, thus allowing removal of
related time.Sleep() calls.