implement unit tests on most daemon functions & return/handle err on ConnectAp()
Refactored daemon into a pkg (daemonPkg) and a command that runs as a daemon.
The pkg now has unit tests. Some functions do not, but these are mostly
covered by existing tests in other packages. As part of this, added Get/Set
functions for various daemon properties with tests.

Modified netman/ConnectAp() to return error on failure to connect, and
server/handlers/ConnectHandler() now checks if error returned and only
starts process to change state to OPERATING if no error.